Rename 'autoGroupTiles' to 'autoMoveTiles'
