+++ packages/scrabble-solver/src/components/Settings/components/AutoGroupTilesSetting/constants.ts
