+++ packages/scrabble-solver/src/components/Settings/components/AutoGroupTilesSetting/index.ts
@@ -1,0 +1,1 @@
+export { default } from './AutoGroupTilesSetting';
