+++ packages/scrabble-solver/src/components/Settings/components/AutoGroupTilesSetting/lib.ts
