+++ packages/scrabble-solver/src/components/Settings/components/AutoGroupTilesSetting/types.ts
