--- packages/scrabble-solver/src/components/Settings/components/AutoMoveTilesSetting/index.ts
@@ -1,1 +1,0 @@
-export { default } from './AutoMoveTilesSetting';
