+++ packages/scrabble-solver/src/components/Settings/components/index.ts
@@ -1,4 +1,4 @@
-export { default as AutoMoveTilesSetting } from './AutoMoveTilesSetting';
+export { default as AutoGroupTilesSetting } from './AutoGroupTilesSetting';
 export { default as ConfigSetting } from './ConfigSetting';
 export { default as LocaleSetting } from './LocaleSetting';
 export { default as OtherSettings } from './OtherSettings';
