+++ packages/scrabble-solver/src/i18n/en.json
@@ -30,10 +30,10 @@
   "settings": "Settings",
   "settings.autoDirectionChange": "Change typing direction on the board with arrow keys",
   "settings.autoDirectionChangeTitle": "Navigating with arrow keys on the board will also change typing direction",
-  "settings.autoMoveTiles": "Group remaining tiles",
-  "settings.autoMoveTiles.left": "On the left",
-  "settings.autoMoveTiles.right": "On the right",
-  "settings.autoMoveTiles.null": "Do not group",
+  "settings.autoGroupTiles": "Group remaining tiles",
+  "settings.autoGroupTiles.left": "On the left",
+  "settings.autoGroupTiles.right": "On the right",
+  "settings.autoGroupTiles.null": "Do not group",
   "settings.game": "Game",
   "settings.language": "Language",
   "settings.other": "Other",
