+++ packages/scrabble-solver/src/i18n/pl.json
@@ -30,10 +30,10 @@
   "settings": "Ustawienia",
   "settings.autoDirectionChange": "Zmieniaj kierunek pisania na planszy strzałkami",
   "settings.autoDirectionChangeTitle": "Nawigacja klawiszami strzałek na planszy będzie również zmieniać kierunek pisania",
-  "settings.autoMoveTiles": "Grupuj pozostałe płytki",
-  "settings.autoMoveTiles.left": "Po lewej",
-  "settings.autoMoveTiles.right": "Po prawej",
-  "settings.autoMoveTiles.null": "Nie grupuj",
+  "settings.autoGroupTiles": "Grupuj pozostałe płytki",
+  "settings.autoGroupTiles.left": "Po lewej",
+  "settings.autoGroupTiles.right": "Po prawej",
+  "settings.autoGroupTiles.null": "Nie grupuj",
   "settings.game": "Gra",
   "settings.language": "Język",
   "settings.other": "Inne",
